requirements.txt
# Dependencies list
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
flask-wtf>=1.2.0
flask-migrate>=4.0.0
werkzeug>=3.0.0
wtforms>=3.1.0
python-dotenv>=1.0.0